_ctypes: extract into module/_ctypes/ following the _socket pattern
importing.rs shrinks by 315 lines.
